backend/requirements.txt
# PhishVision Backend Dependencies

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database
flask-sqlalchemy>=3.1.0

# Environment
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# HTML Parsing (for website cloning)
beautifulsoup4>=4.12.0

# QR Code Generation
qrcode[pil]>=7.4.0
pillow>=10.0.0

# AI Template Generation
google-generativeai>=0.3.0

# DNS/Email Validation
dnspython>=2.4.0

# SMS Integration
twilio>=8.10.0
vonage>=4.7.0

# Document Generation
xhtml2pdf>=0.2.11
